Move into Shared SqlErrorCollection.cs (#1305)
